Calculator operation activities now return results. tests needed

+++ app/src/main/java/com/rave/simplemath/model/remote/MathService.kt
@@ -11,5 +11,5 @@ import retrofit2.http.Query
 interface MathService {
 
     @GET(".")
-    suspend fun evaluateExpression(@Query("expr", encoded = false) expr: String): Int
+    suspend fun evaluateExpression(@Query("expr", encoded = false) expr: String): Double
 }
